backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Google Gemini AI
google-generativeai==0.8.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
Pillow>=10.3.0
pytesseract==0.3.13

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Utilities
aiofiles==24.1.0
youtube-transcript-api==0.6.2


